refactor: auto-split test_api_architecture.py
