+++ tests/config/test_launcher_manifest.py.bak
